+++ demisto_sdk/commands/pre_commit/tests/pre_commit_test.py
@@ -2,7 +2,7 @@ import itertools
 import os
 from dataclasses import dataclass
 from pathlib import Path
-from typing import List, Optional
+from typing import Dict, List, Optional
 
 import pytest
 
@@ -10,6 +10,7 @@ import demisto_sdk.commands.pre_commit.pre_commit_command as pre_commit_command
 import demisto_sdk.commands.pre_commit.pre_commit_context as context
 from demisto_sdk.commands.common.handlers import DEFAULT_YAML_HANDLER as yaml
 from demisto_sdk.commands.common.legacy_git_tools import git_path
+from demisto_sdk.commands.common.native_image import NativeImageConfig
 from demisto_sdk.commands.pre_commit.hooks.docker import DockerHook
 from demisto_sdk.commands.pre_commit.hooks.hook import Hook, join_files
 from demisto_sdk.commands.pre_commit.hooks.mypy import MypyHook
@@ -25,6 +26,7 @@ from demisto_sdk.commands.pre_commit.pre_commit_command import (
     subprocess,
 )
 from TestSuite.repo import Repo
+from TestSuite.test_tools import ChangeCWD
 
 TEST_DATA_PATH = (
     Path(git_path()) / "demisto_sdk" / "commands" / "pre_commit" / "tests" / "test_data"
@@ -35,6 +37,17 @@ PYTHON_VERSION_TO_FILES = {
     "3.9": {Path("Packs/Pack1/Integrations/integration2/integration2.py")},
     "3.10": {Path("Packs/Pack1/Integrations/integration3/integration3.py")},
 }
+
+
+@pytest.fixture()
+def native_image_config(mocker, repo) -> NativeImageConfig:
+    native_image_config = NativeImageConfig.from_path(
+        repo.docker_native_image_config.path
+    )
+    mocker.patch.object(
+        NativeImageConfig, "get_instance", return_value=native_image_config
+    )
+    return native_image_config
 
 
 @dataclass(frozen=True)
@@ -85,23 +98,30 @@ class MockProcess:
         return "", ""
 
 
-def test_config_files(mocker, repo: Repo):
+def test_config_files(mocker, repo: Repo, native_image_config):
     """
     Given:
-        A repository with different scripts and integration of different python versions
+        A repository with different scripts and integration of different python versions which require
+        split hooks for pylint
 
     When:
         Calling demisto-sdk pre-commit
 
     Then:
-        Categorize the scripts and integration by python version, and make sure that pre-commit configuration is created for each
+        - Categorize the scripts and integration by python version,
+          and make sure that pre-commit configuration is created
+        - make sure split hooks are created properly
+        - make sure that the created hooks are pylint based only as its the only hook that should be split
     """
-    mocker.patch.object(DockerHook, "__init__", return_value=None)
-    mocker.patch.object(
-        DockerHook,
-        "prepare_hook",
-        return_value=[{"id": "run-in-docker"}],
+
+    def devtest_side_effect(image_tag: str, is_powershell: bool, should_pull: bool):
+        return image_tag
+
+    mocker.patch(
+        "demisto_sdk.commands.pre_commit.hooks.docker.devtest_image",
+        side_effect=devtest_side_effect,
     )
+
     mocker.patch.object(
         context,
         "PRECOMMIT_TEMPLATE_PATH",
@@ -114,16 +134,6 @@ def test_config_files(mocker, repo: Repo):
         pre_commit_command,
         "PRECOMMIT_CONFIG_MAIN_PATH",
         Path(repo.path) / ".pre-commit-config.yaml",
-    )
-    mocker.patch.object(
-        context,
-        "PRECOMMIT_DOCKER_CONFIGS",
-        Path(repo.path) / "docker-config",
-    )
-    mocker.patch.object(
-        pre_commit_command,
-        "PRECOMMIT_DOCKER_CONFIGS",
-        Path(repo.path) / "docker-config",
     )
     mocker.patch.object(context, "PRECOMMIT_CONFIG", Path(repo.path) / "config")
     integration1 = pack1.create_integration(
@@ -161,7 +171,13 @@ def test_config_files(mocker, repo: Repo):
 
     python_version_to_files, _ = group_by_language(files_to_run)
     pre_commit_context = pre_commit_command.PreCommitContext(
-        None, None, None, python_version_to_files, ""
+        None,
+        None,
+        None,
+        python_version_to_files,
+        "",
+        pre_commit_template_path=TEST_DATA_PATH
+        / ".pre-commit-config_template-test.yaml",
     )
     assert (
         Path(script1.yml.path).relative_to(repo.path)
@@ -191,8 +207,41 @@ def test_config_files(mocker, repo: Repo):
 
     PreCommitRunner.prepare_and_run(pre_commit_context)
     assert (Path(repo.path) / ".pre-commit-config.yaml").exists()
-    assert list((Path(repo.path) / "docker-config").iterdir())
-    assert (Path(repo.path) / ".pre-commit-config-needs.yaml").exists()
+
+
+def test_handle_api_modules(mocker, git_repo: Repo):
+    """
+    Given:
+        - A repository with a pack that contains an API module and a pack that contains an integration that uses the API module
+
+    When:
+        - Running demisto-sdk pre-commit
+
+    Then:
+        - Ensure that the API module is added to the files to run
+        - Ensure that the integration that uses the API module is added to the files to run, both related to the *integration*
+    """
+    pack1 = git_repo.create_pack("ApiModules")
+    script = pack1.create_script("TestApiModule")
+    pack2 = git_repo.create_pack("Pack2")
+    integration = pack2.create_integration(
+        "integration1", code="from TestApiModule import *"
+    )
+    mocker.patch.object(pre_commit_command, "CONTENT_PATH", Path(git_repo.path))
+    with ChangeCWD(git_repo.path):
+        git_repo.create_graph()
+        files_to_run = group_by_language(
+            {Path(script.yml.path).relative_to(git_repo.path)}
+        )
+    files_to_run = {(path, obj.path) for path, obj in files_to_run[0]["2.7"]}
+    assert (
+        Path(script.yml.path).relative_to(git_repo.path),
+        integration.object.path.relative_to(git_repo.path),
+    ) in files_to_run
+    assert (
+        Path(integration.yml.path).relative_to(git_repo.path),
+        integration.object.path.relative_to(git_repo.path),
+    ) in files_to_run
 
 
 def test_mypy_hooks(mocker):
@@ -207,7 +256,8 @@ def test_mypy_hooks(mocker):
             "--follow-imports=silent",
             "--allow-redefinition",
             "--python-version=3.10",
-        ]
+        ],
+        "id": "mypy",
     }
     mocker.patch.object(
         PreCommitContext, "python_version_to_files", PYTHON_VERSION_TO_FILES
@@ -232,7 +282,11 @@ def test_ruff_hook(github_actions, mocker):
         PreCommitContext, "python_version_to_files", PYTHON_VERSION_TO_FILES
     )
     ruff_hook = create_hook(
-        {"args": ["--fix"], "args:nightly": ["--config=nightly_ruff.toml"]}
+        {
+            "args": ["--fix"],
+            "args:nightly": ["--config=nightly_ruff.toml"],
+            "id": "ruff",
+        }
     )
 
     mocker.patch.dict(
@@ -261,7 +315,11 @@ def test_ruff_hook_nightly_mode(mocker):
         PreCommitContext, "python_version_to_files", PYTHON_VERSION_TO_FILES
     )
     ruff_hook = create_hook(
-        {"args": ["--fix"], "args:nightly": ["--config=nightly_ruff.toml"]},
+        {
+            "args": ["--fix"],
+            "args:nightly": ["--config=nightly_ruff.toml"],
+            "id": "ruff",
+        },
         mode="nightly",
     )
 
@@ -279,7 +337,9 @@ def test_validate_format_hook_nightly_mode_and_all_files(mocker):
     """
     Testing validate_format hook created successfully (the -a flag is added and the -i arg is not exist)
     """
-    validate_format_hook = create_hook({"args": []}, mode="nightly", all_files=True)
+    validate_format_hook = create_hook(
+        {"args": [], "id": "validate"}, mode="nightly", all_files=True
+    )
     mocker.patch.object(
         PreCommitContext, "python_version_to_files", PYTHON_VERSION_TO_FILES
     )
@@ -296,7 +356,7 @@ def test_validate_format_hook_nightly_mode(mocker):
     Testing validate_format hook created successfully (the -i arg is added and the -a flag is not exist, even in nightly mode)
     """
     validate_format_hook = create_hook(
-        {"args": []}, mode="nightly", input_files=[Path("file1.py")]
+        {"args": [], "id": "validate"}, mode="nightly", input_files=[Path("file1.py")]
     )
     mocker.patch.object(
         PreCommitContext, "python_version_to_files", PYTHON_VERSION_TO_FILES
@@ -313,7 +373,7 @@ def test_validate_format_hook_all_files(mocker):
     """
     Testing validate_format hook created successfully (the -i arg is added and the -a flag is not exist)
     """
-    validate_format_hook = create_hook({"args": []}, all_files=True)
+    validate_format_hook = create_hook({"args": [], "id": "validate"}, all_files=True)
     mocker.patch.object(
         PreCommitContext, "python_version_to_files", PYTHON_VERSION_TO_FILES
     )
@@ -355,6 +415,7 @@ class TestPreprocessFiles:
         expected_output = {
             Path(integration.yml.rel_path),
             Path(integration.code.rel_path),
+            Path(integration.test.rel_path),
         }
         mocker.patch.object(GitUtil, "get_all_files", return_value=relative_paths)
         output = preprocess_files(input_files=input_files)
@@ -437,7 +498,13 @@ def test_exclude_hooks_by_version(mocker, repo: Repo):
         "3.8": {(Path("file2.py"), None)},
     }
     pre_commit_context = pre_commit_command.PreCommitContext(
-        None, None, None, python_version_to_files, ""
+        None,
+        None,
+        None,
+        python_version_to_files,
+        "",
+        pre_commit_template_path=TEST_DATA_PATH
+        / ".pre-commit-config_template-test.yaml",
     )
     PreCommitRunner.prepare_hooks(pre_commit_context)
 
@@ -468,7 +535,13 @@ def test_exclude_hooks_by_support_level(mocker, repo: Repo):
         "3.8": {(Path("file2.py"), Obj(support_level="community"))},
     }
     pre_commit_context = pre_commit_command.PreCommitContext(
-        None, None, None, python_version_to_files, ""
+        None,
+        None,
+        None,
+        python_version_to_files,
+        "",
+        pre_commit_template_path=TEST_DATA_PATH
+        / ".pre-commit-config_template-test.yaml",
     )
 
     PreCommitRunner.prepare_hooks(pre_commit_context)
@@ -514,7 +587,7 @@ def test_coverage_analyze_general_hook(mode, expected_args):
     """
 
     coverage_analyze_hook = create_hook(
-        {"args": args, "args:nightly": args_nightly},
+        {"args": args, "args:nightly": args_nightly, "id": "coverage-analyze"},
         mode=mode,
         all_files=True,
         input_files=[Path("file1.py")],
@@ -527,15 +600,16 @@ def test_coverage_analyze_general_hook(mode, expected_args):
 @pytest.mark.parametrize(
     "hook, expected_result",
     [
-        ({"files": r"\.py$", "exclude": r"_test\.py$"}, ["file1.py", "file6.py"]),
         (
-            {
-                "files": r"\.py$",
-            },
+            {"files": r"\.py$", "exclude": r"_test\.py$", "id": "test"},
+            ["file1.py", "file6.py"],
+        ),
+        (
+            {"files": r"\.py$", "id": "test"},
             ["file1.py", "file6.py", "file2_test.py"],
         ),
         (
-            {},
+            {"id": "test"},
             [
                 "file1.py",
                 "file2_test.py",
@@ -545,7 +619,7 @@ def test_coverage_analyze_general_hook(mode, expected_args):
                 "file6.py",
             ],
         ),
-        ({"files": r"\.ps1$"}, ["file3.ps1"]),
+        ({"files": r"\.ps1$", "id": "test"}, ["file3.ps1"]),
     ],
 )
 def test_filter_files_matching_hook_config(hook, expected_result):
@@ -588,6 +662,13 @@ def test_skip_hook_with_mode(mocker):
     Then:
         Don't generate the skipped hooks
     """
+
+    def get_repos(_pre_commit_config: Dict) -> Dict:
+        repos = {}
+        for repo in _pre_commit_config["repos"]:
+            repos[repo["repo"]] = repo
+        return repos
+
     mocker.patch.object(
         context,
         "PRECOMMIT_TEMPLATE_PATH",
@@ -598,9 +679,15 @@ def test_skip_hook_with_mode(mocker):
         "3.8": {(Path("file2.py"), None)},
     }
     pre_commit_runner = pre_commit_command.PreCommitContext(
-        None, None, "nightly", python_version_to_files, ""
+        None,
+        None,
+        "nightly",
+        python_version_to_files,
+        "",
+        pre_commit_template_path=TEST_DATA_PATH
+        / ".pre-commit-config_template-test.yaml",
     )
-    repos = pre_commit_runner._get_repos(pre_commit_runner.precommit_template)
+    repos = get_repos(pre_commit_runner.precommit_template)
     assert not repos["https://github.com/charliermarsh/ruff-pre-commit"]["hooks"]
     assert "is-gitlab-changed" not in {
         hook.get("id") for hook in repos["local"]["hooks"]
@@ -626,7 +713,7 @@ def test_system_hooks():
 
     Path(sys.executable).parent
     system_hook = create_hook(
-        {"args": [], "entry": "demisto-sdk", "language": "system"}
+        {"args": [], "entry": "demisto-sdk", "language": "system", "id": "test"}
     )
     SystemHook(**system_hook).prepare_hook()
     assert (