+++ demisto_sdk/commands/validate/initializer.py
@@ -24,12 +24,12 @@ from demisto_sdk.commands.common.constants import (
     PathLevel,
 )
 from demisto_sdk.commands.common.content import Content
-from demisto_sdk.commands.common.content_constant_paths import CONTENT_PATH
 from demisto_sdk.commands.common.logger import logger
 from demisto_sdk.commands.common.tools import (
     detect_file_level,
     get_file_by_status,
     get_relative_path_from_packs_dir,
+    is_external_repo,
     specify_files_from_directory,
 )
 from demisto_sdk.commands.content_graph.objects.base_content import BaseContent
@@ -179,9 +179,12 @@ class Initializer:
             self.committed_only = True
 
         elif self.branch_name in ["master", "main", DEMISTO_GIT_PRIMARY_BRANCH]:
-            raise Exception(
-                "Running on master branch while using git is ill advised.\nrun: 'git checkout -b NEW_BRANCH_NAME' and rerun the command."
-            )
+            self.git_util
+            message = "Running on master branch while using git is ill advised.\nrun: 'git checkout -b NEW_BRANCH_NAME' and rerun the command."
+            if not is_external_repo() or self.committed_only:
+                logger.warning(message)
+            else:
+                raise Exception(message)
 
     def print_git_config(self):
         """Printing the git configurations - all the relevant flags."""
@@ -323,21 +326,25 @@ class Initializer:
                 set(self.load_files(self.file_path.split(",")))
             )
         elif self.all_files:
-            content_dto = ContentDTO.from_path(CONTENT_PATH)
+            logger.info("Running validation on all files.")
+            content_dto = ContentDTO.from_path()
             if not isinstance(content_dto, ContentDTO):
                 raise Exception("no content found")
             content_objects_to_run = set(content_dto.packs)
         else:
-            self.use_git = (True,)
+            self.use_git = True
             self.committed_only = True
             (
                 content_objects_to_run,
                 invalid_content_items,
                 non_content_items,
             ) = self.get_files_using_git()
-        content_objects_to_run_with_packs: Set[BaseContent] = self.get_items_from_packs(
-            content_objects_to_run
-        )
+        if not self.use_git:
+            content_objects_to_run_with_packs: Set[
+                BaseContent
+            ] = self.get_items_from_packs(content_objects_to_run)
+        else:
+            content_objects_to_run_with_packs = content_objects_to_run
         for non_content_item in non_content_items:
             logger.warning(
                 f"Invalid content path provided: {str(non_content_item)}. Please provide a valid content item or pack path."
@@ -483,9 +490,15 @@ class Initializer:
                     obj.git_status = git_status
                     # Check if the file exists
                     if git_status in (GitStatuses.MODIFIED, GitStatuses.RENAMED):
-                        obj.old_base_content_object = BaseContent.from_path(
-                            old_path, git_sha=git_sha, raise_on_exception=True
-                        )
+                        try:
+                            obj.old_base_content_object = BaseContent.from_path(
+                                old_path, git_sha=git_sha, raise_on_exception=True
+                            )
+                        except (NotAContentItemException, InvalidContentItemException):
+                            logger.debug(
+                                f"Could not parse the old_base_content_object for {obj.path}, setting a copy of the object as the old_base_content_object."
+                            )
+                            obj.old_base_content_object = obj.copy(deep=True)
                     else:
                         obj.old_base_content_object = obj.copy(deep=True)
                     if obj.old_base_content_object:
@@ -507,7 +520,7 @@ class Initializer:
             path_str = str(path)
             if self.is_unrelated_path(path_str):
                 continue
-            if "Integrations" in path_str or "Scripts" in path_str:
+            if f"/{INTEGRATIONS_DIR}/" in path_str or f"/{SCRIPTS_DIR}/" in path_str:
                 if path_str.endswith(".yml"):
                     statuses_dict[path] = git_status
                 elif self.is_code_file(path, path_str):
@@ -522,7 +535,7 @@ class Initializer:
                     path = Path(path.parent / f"{path.parts[-2]}.yml")
                     if path not in statuses_dict:
                         statuses_dict[path] = None
-            elif "Playbooks" in path_str:
+            elif f"/{PLAYBOOKS_DIR }/" in path_str:
                 if path_str.endswith(".yml"):
                     statuses_dict[path] = git_status
                 else:
@@ -584,7 +597,7 @@ class Initializer:
             path_str = str(path)
             if self.is_unrelated_path(path_str):
                 continue
-            if INTEGRATIONS_DIR in path_str or SCRIPTS_DIR in path_str:
+            if f"/{INTEGRATIONS_DIR}/" in path_str or f"/{SCRIPTS_DIR}/" in path_str:
                 if path_str.endswith(".yml"):
                     paths_set.add(path)
                 elif self.is_code_file(path, path_str):
@@ -594,7 +607,7 @@ class Initializer:
                     paths_set.add(path)
                 else:
                     paths_set.add(path.parent / f"{path.parts[-2]}.yml")
-            elif PLAYBOOKS_DIR in path_str:
+            elif f"/{PLAYBOOKS_DIR }/" in path_str:
                 if path_str.endswith(".yml"):
                     paths_set.add(path)
                 else: