+++ demisto_sdk/commands/validate/validators/BA_validators/BA100_is_valid_version.py
@@ -20,7 +20,6 @@ from demisto_sdk.commands.content_graph.objects.list import List as List_Obj
 from demisto_sdk.commands.content_graph.objects.mapper import Mapper
 from demisto_sdk.commands.content_graph.objects.playbook import Playbook
 from demisto_sdk.commands.content_graph.objects.script import Script
-from demisto_sdk.commands.content_graph.objects.test_playbook import TestPlaybook
 from demisto_sdk.commands.content_graph.objects.widget import Widget
 from demisto_sdk.commands.content_graph.objects.wizard import Wizard
 from demisto_sdk.commands.validate.validators.base_validator import (
@@ -42,7 +41,6 @@ ContentTypes = Union[
     IncidentField,
     IndicatorField,
     IndicatorType,
-    TestPlaybook,
     GenericDefinition,
     GenericField,
     GenericModule,
@@ -54,7 +52,10 @@ ContentTypes = Union[
 
 class IsValidVersionValidator(BaseValidator[ContentTypes]):
     error_code = "BA100"
-    description = "Validate that the content item version equal -1."
+    description = "Marketplace content set to -1 makes it easier to tell from modified, versioned content."
+    rationale = (
+        "The version for system content items should always be -1 as per the standard."
+    )
     error_message = (
         "The version for our files should always be -1, please update the file."
     )