+++ demisto_sdk/scripts/add_marketplace_to_schemas.py
@@ -8,9 +8,11 @@ from demisto_sdk.commands.common.tools import get_file, write_dict
 
 GIT_ROOT = Path(git_path())
 SCHEMA_FOLDER = GIT_ROOT / "demisto_sdk" / "commands" / "common" / "schemas"
-NON_SUPPORTED_KEYS = ["id"]
+NON_SUPPORTED_KEYS = [""]
 SUPPORTED_KEYS = [
+    "id",
     "isfetch",
+    "isfetchevents",
     "defaultValue",
     "defaultvalue",
     "required",